requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.13.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
executing==2.1.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.54.1
fsspec==2024.10.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
gradio==5.7.1
gradio_client==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
orjson==3.10.12
packaging==24.1
pandas==2.2.3
parsivar==0.2.3.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.13
pyparsing==3.2.0
pyreadline3==3.5.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.65
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruff==0.8.1
safehttpx==0.1.6
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.38.0
streamlit-cropper==0.2.2
streamlit-image-coordinates==0.2.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2025.3.13
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.27.1
watchdog==4.0.2
wcwidth==0.2.13
websockets==12.0
xyzservices==2025.1.0
